refactor: simplify UI repository and helper view models
