+++ src/VaultSync.UI/Infrastructure/DetachedTask.cs
@@ -5,6 +5,26 @@ namespace VaultSync.UI.Infrastructure;
 
 public static class DetachedTask
 {
+    public static void Run(Action operation, string operationName)
+    {
+        _ = Task.Run(() =>
+        {
+            try
+            {
+                operation();
+            }
+            catch (Exception ex)
+            {
+                LogFailure(operationName, ex);
+            }
+        });
+    }
+
+    public static void Run(Func<Task> operation, string operationName)
+    {
+        _ = Task.Run(() => RunAsync(operation, operationName));
+    }
+
     public static async Task RunAsync(Func<Task> operation, string operationName)
     {
         try
@@ -13,8 +33,13 @@ public static class DetachedTask
         }
         catch (Exception ex)
         {
-            DiagnosticsLogger.Record(
-                $"Detached operation failed ({operationName}): {ex.GetType().Name} - {ex.Message}");
+            LogFailure(operationName, ex);
         }
+    }
+
+    private static void LogFailure(string operationName, Exception ex)
+    {
+        DiagnosticsLogger.Record(
+            $"Detached operation failed ({operationName}): {ex.GetType().Name} - {ex.Message}");
     }
 }
