+++ src/VaultSync.UI/Services/IRepositoryFactory.cs
@@ -1,0 +1,11 @@
+using VaultSync.Core.Config;
+using VaultSync.Core.Repositories;
+
+namespace VaultSync.UI.Services;
+
+public interface IRepositoryFactory
+{
+    SqliteRepository Create(AppConfig? config = null);
+
+    string ResolveDbPath(AppConfig? config = null);
+}
