+++ src/VaultSync.UI/ViewModels/SettingsViewModel.cs
@@ -126,6 +126,7 @@ namespace VaultSync.UI
         private bool _showRsyncStatusHint;
         private string _selectedLanguageCode = "en";
         private readonly IAppConfigStore _configStore;
+        private readonly IRepositoryFactory _repositoryFactory;
         private readonly CredentialVault _credentialVault = CredentialVault.Instance;
         private readonly BackupEncryptionSecretService _backupEncryptionSecretService = new();
         private readonly NetworkMountService _networkMountService = new();
@@ -447,10 +448,11 @@ namespace VaultSync.UI
             ShowLegacyBackupLocation = !UseAdvancedDestinations;
         }
 
-        public SettingsViewModel(LocalizationService localizationService, IAppConfigStore? configStore = null)
+        public SettingsViewModel(LocalizationService localizationService, IAppConfigStore? configStore = null, IRepositoryFactory? repositoryFactory = null)
         {
             _localizationService = localizationService;
             _configStore = configStore ?? StaticAppConfigStore.Instance;
+            _repositoryFactory = repositoryFactory ?? new SqliteRepositoryFactory(_configStore);
             _selectedLanguageCode = localizationService.CurrentLanguage;
             _localizationService.LanguageChanged += () =>
             {
@@ -2245,19 +2247,12 @@ namespace VaultSync.UI
 
         private void PersistLanguage()
         {
-            _ = Task.Run(() =>
+            DetachedTask.Run(() =>
             {
-                try
-                {
-                    AppConfig cfg = _configStore.Load();
-                    cfg.Advanced.Language = _selectedLanguageCode;
-                    _configStore.Save(cfg);
-                }
-                catch
-                {
-                    // Best effort; avoid crashing when persisting language.
-                }
-            });
+                AppConfig cfg = _configStore.Load();
+                cfg.Advanced.Language = _selectedLanguageCode;
+                _configStore.Save(cfg);
+            }, nameof(PersistLanguage));
         }
 
         public void UpdateUpdateCheckStatus(DateTimeOffset? lastCheck, string? errorMessage)
@@ -3202,7 +3197,7 @@ namespace VaultSync.UI
                     // Dev helper: reset the VaultSync SQLite DB to a "fresh install" state
                     // without touching any real project files or backup folders on disk.
                     AppConfig cfg  = _configStore.Load();
-                    var repo = new SqliteRepository(cfg.DbPath ?? string.Empty);
+                    var repo = _repositoryFactory.Create(cfg);
 
                     repo.EnsureSchema();
                     repo.ResetAllData();
@@ -3374,22 +3369,7 @@ namespace VaultSync.UI
                 NotificationSeverity.Info,
                 L("Settings.Advanced.TelemetryExportTitle", "Telemetry export"));
 
-            try
-            {
-                string? folder = Path.GetDirectoryName(result.ZipPath);
-                if (!string.IsNullOrWhiteSpace(folder))
-                {
-                    Process.Start(new ProcessStartInfo
-                    {
-                        FileName = folder,
-                        UseShellExecute = true
-                    });
-                }
-            }
-            catch
-            {
-                // best-effort
-            }
+            TryOpenContainingFolder(result.ZipPath);
         }
 
         private void OpenLogConsole()
@@ -3481,9 +3461,14 @@ namespace VaultSync.UI
                 NotificationSeverity.Info,
                 L("Settings.Advanced.SupportBundle", "Support bundle"));
 
+            TryOpenContainingFolder(result.ZipPath);
+        }
+
+        private static void TryOpenContainingFolder(string? artifactPath)
+        {
             try
             {
-                string? folder = Path.GetDirectoryName(result.ZipPath);
+                string? folder = Path.GetDirectoryName(artifactPath);
                 if (!string.IsNullOrWhiteSpace(folder))
                 {
                     Process.Start(new ProcessStartInfo
@@ -3519,7 +3504,7 @@ namespace VaultSync.UI
                 BackupRetentionSimulationResult result = await Task.Run(() =>
                 {
                     AppConfig cfg = _configStore.Load();
-                    var repo = new SqliteRepository(cfg.DbPath ?? string.Empty);
+                    var repo = _repositoryFactory.Create(cfg);
                     var service = new BackupRetentionSimulationService(repo);
                     return service.Simulate(ClampInt(cfg.Backups.MaxSnapshotsPerProject, 1, 999, 20));
                 }).ConfigureAwait(false);
@@ -3647,7 +3632,7 @@ namespace VaultSync.UI
                 BackupIndexRepairPlan plan = await Task.Run(() =>
                 {
                     AppConfig cfg = _configStore.Load();
-                    var repo = new SqliteRepository(cfg.DbPath ?? string.Empty);
+                    var repo = _repositoryFactory.Create(cfg);
                     var service = new BackupIndexRepairService(repo);
                     return service.BuildPlan();
                 }).ConfigureAwait(false);
@@ -3722,7 +3707,7 @@ namespace VaultSync.UI
                 int applied = await Task.Run(() =>
                 {
                     AppConfig cfg = _configStore.Load();
-                    var repo = new SqliteRepository(cfg.DbPath ?? string.Empty);
+                    var repo = _repositoryFactory.Create(cfg);
                     var service = new BackupIndexRepairService(repo);
                     return service.ApplyPlan(plan);
                 }).ConfigureAwait(false);
@@ -3883,7 +3868,7 @@ namespace VaultSync.UI
                 await Task.Run(() =>
                 {
                     AppConfig cfg = _configStore.Load();
-                    var repo = new SqliteRepository(cfg.DbPath ?? string.Empty);
+                    var repo = _repositoryFactory.Create(cfg);
                     repo.UpdateProjectPreferredDestination(item.ProjectId, EmptyToNull(item.ImportedPreferredDestinationId));
                     repo.UpdateProjectRestoreMode(item.ProjectId, EmptyToNull(item.ImportedRestoreMode));
                     repo.UpdateProjectVerificationPolicy(item.ProjectId, EmptyToNull(item.ImportedVerificationPolicy));
@@ -4408,197 +4393,4 @@ namespace VaultSync.UI
 
     }
 
-    public class BackupDestinationViewModel : VaultSync.UI.ViewModels.ViewModelBase
-    {
-        private string _alias = string.Empty;
-        public string Alias
-        {
-            get => _alias;
-            set
-            {
-                if (SetField(ref _alias, value))
-                {
-                    OnPropertyChanged(nameof(DisplayName));
-                }
-            }
-        }
-
-        private string _path = string.Empty;
-        public string Path
-        {
-            get => _path;
-            set
-            {
-                if (SetField(ref _path, value))
-                {
-                    OnPropertyChanged(nameof(DisplayName));
-                }
-            }
-        }
-
-        private NetworkCredentialViewModel? _selectedCredential;
-        public NetworkCredentialViewModel? SelectedCredential
-        {
-            get => _selectedCredential;
-            set
-            {
-                if (SetField(ref _selectedCredential, value))
-                {
-                    CredentialName = value?.Name ?? string.Empty;
-                    OnPropertyChanged(nameof(NeedsCredentialWarning));
-                }
-            }
-        }
-
-        private string _credentialName = string.Empty;
-        public string CredentialName
-        {
-            get => _credentialName;
-            set
-            {
-                if (SetField(ref _credentialName, value))
-                {
-                    // Keep SelectedCredential in sync when only the name changes
-                    if (SelectedCredential is null || !string.Equals(SelectedCredential.Name, value, StringComparison.OrdinalIgnoreCase))
-                    {
-                        // Selection will be resolved via SettingsViewModel handler
-                    }
-                }
-            }
-        }
-
-        // Used by the Settings UI ComboBox. When the Settings page is unloaded,
-        // Avalonia may temporarily clear SelectedItem and push a null/empty value
-        // back into the binding; ignore that so the destination keeps its selection.
-        public string SelectedCredentialName
-        {
-            get => CredentialName ?? string.Empty;
-            set
-            {
-                if (string.IsNullOrWhiteSpace(value))
-                    return;
-
-                CredentialName = value;
-            }
-        }
-
-        private bool _active = true;
-        public bool Active { get => _active; set => SetField(ref _active, value); }
-
-        private bool _autoMount;
-        public bool AutoMount
-        {
-            get => _autoMount;
-            set
-            {
-                if (SetField(ref _autoMount, value))
-                {
-                    OnPropertyChanged(nameof(NeedsCredentialWarning));
-                }
-            }
-        }
-
-        private bool _autoUnmount;
-        public bool AutoUnmount { get => _autoUnmount; set => SetField(ref _autoUnmount, value); }
-
-        private bool _preMounted = true;
-        public bool PreMounted
-        {
-            get => _preMounted;
-            set
-            {
-                if (SetField(ref _preMounted, value))
-                {
-                    OnPropertyChanged(nameof(NeedsCredentialWarning));
-                }
-            }
-        }
-
-        public bool NeedsCredentialWarning =>
-            AutoMount && !PreMounted && SelectedCredential is null;
-
-        private bool _enableMetadataSync = true;
-        public bool EnableMetadataSync { get => _enableMetadataSync; set => SetField(ref _enableMetadataSync, value); }
-
-        private bool _autoImportMetadata = true;
-        public bool AutoImportMetadata { get => _autoImportMetadata; set => SetField(ref _autoImportMetadata, value); }
-
-        private bool _forceMetadataBackfill;
-        public bool ForceMetadataBackfill { get => _forceMetadataBackfill; set => SetField(ref _forceMetadataBackfill, value); }
-
-        private int _retryMaxAttempts = 1;
-        public int RetryMaxAttempts
-        {
-            get => _retryMaxAttempts;
-            set => SetField(ref _retryMaxAttempts, Math.Clamp(value, 1, 10));
-        }
-
-        private int _retryBackoffSeconds = 10;
-        public int RetryBackoffSeconds
-        {
-            get => _retryBackoffSeconds;
-            set => SetField(ref _retryBackoffSeconds, Math.Clamp(value, 1, 300));
-        }
-
-        private bool _enableCheckpointResume = true;
-        public bool EnableCheckpointResume
-        {
-            get => _enableCheckpointResume;
-            set => SetField(ref _enableCheckpointResume, value);
-        }
-
-        private double _softQuotaGb;
-        public double SoftQuotaGb
-        {
-            get => _softQuotaGb;
-            set => SetField(ref _softQuotaGb, Math.Clamp(value, 0d, 1024d * 1024d));
-        }
-
-        private int _quotaWarningPercent = 85;
-        public int QuotaWarningPercent
-        {
-            get => _quotaWarningPercent;
-            set => SetField(ref _quotaWarningPercent, Math.Clamp(value, 50, 99));
-        }
-
-        private string _lastTestStatus = string.Empty;
-        public string LastTestStatus
-        {
-            get => _lastTestStatus;
-            set => SetField(ref _lastTestStatus, value);
-        }
-
-        private string _lastTestSeverity = "Info";
-        public string LastTestSeverity
-        {
-            get => _lastTestSeverity;
-            set => SetField(ref _lastTestSeverity, value);
-        }
-
-        public string DisplayName => string.IsNullOrWhiteSpace(Alias) ? Path : Alias;
-    }
-
-    public class NetworkCredentialViewModel : VaultSync.UI.ViewModels.ViewModelBase
-    {
-        private string _name = string.Empty;
-        public string Name { get => _name; set => SetField(ref _name, value); }
-
-        private string _username = string.Empty;
-        public string Username { get => _username; set => SetField(ref _username, value); }
-
-        private string _domain = string.Empty;
-        public string Domain { get => _domain; set => SetField(ref _domain, value); }
-
-        private string _keyRef = string.Empty;
-        public string KeyRef { get => _keyRef; set => SetField(ref _keyRef, value); }
-
-        private bool _useKeychain = true;
-        public bool UseKeychain { get => _useKeychain; set => SetField(ref _useKeychain, value); }
-
-        private string _password = string.Empty;
-        public string Password { get => _password; set => SetField(ref _password, value); }
-
-        private bool _showPassword = false;
-        public bool ShowPassword { get => _showPassword; set => SetField(ref _showPassword, value); }
-    }
 }
